Create Functions for Each of The Butto
I've created or should I say refactor the function playRound(). This is how the function works now. In the previous code all the javascript code was inside the index.html but now moved it to a separate file.
The function playRound is now divided into three a parts one for each button: rock, paper, scissors.
I used the DOM event listeners to play the game this time instead of asking the user to type their choices.